requirements.txt
# Core AI & Image Processing
tensorflow-cpu>=2.16.0
keras>=3.0.0
tensorflow-datasets>=4.9.4
numpy>=1.26.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# UI & Visualizationgit status
gradio>=5.23.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Networking & API (USDA)
requests>=2.31.0
python-dotenv>=1.0.0

# Critical Fix for Python 3.13 (Gradio/Pydub support)
audioop-lts; python_version >= "3.13"
